Split `ProvisioningAccountTest` class
