+++ tests/Integration/Provisioning/ProvisioningIntegrationTestCase.php
@@ -14,6 +14,7 @@ use Cloudinary\Api\Provisioning\AccountApi;
 use Cloudinary\Exception\ConfigurationException;
 use Cloudinary\Test\CloudinaryTestCase;
 use Exception;
+use PHPUnit_Framework_Constraint_IsType as IsType;
 use RuntimeException;
 
 /**
@@ -33,8 +34,8 @@ abstract class ProvisioningIntegrationTestCase extends CloudinaryTestCase
         try {
             self::$accountApi = new AccountApi();
         } catch (ConfigurationException $ce) {
-            static::$skipAllTests = true;
-            static::$skipReason  = $ce->getMessage();
+            self::$skipAllTests = true;
+            self::$skipReason  = $ce->getMessage();
         }
     }
 
@@ -42,11 +43,10 @@ abstract class ProvisioningIntegrationTestCase extends CloudinaryTestCase
     {
         parent::setUp();
 
-        if (static::$skipAllTests === true) {
-            $this->markTestSkipped("Skipping provisioning API test, reason: " . static::$skipReason);
+        if (self::$skipAllTests === true) {
+            self::markTestSkipped("Skipping provisioning API test, reason: " . static::$skipReason);
         }
     }
-
 
     /**
      * Delete an account user
@@ -124,5 +124,157 @@ abstract class ProvisioningIntegrationTestCase extends CloudinaryTestCase
         } catch (Exception $e) {
             //@TODO: Use logger to print ERROR message
         }
+    }
+
+    /**
+     * Assert that a given object is a valid user account.
+     * Optionally checks it against given values.
+     *
+     * @param mixed $user
+     * @param array $values
+     */
+    protected static function assertValidAccountUser($user, $values = [])
+    {
+        self::assertObjectStructure(
+            $user,
+            [
+                'id' => IsType::TYPE_STRING,
+                'name' => IsType::TYPE_STRING,
+                'role' => IsType::TYPE_STRING,
+                'email' => IsType::TYPE_STRING,
+                'pending' => IsType::TYPE_BOOL,
+                'enabled' => IsType::TYPE_BOOL,
+                'created_at' => IsType::TYPE_STRING,
+                'all_sub_accounts' => IsType::TYPE_BOOL
+            ]
+        );
+
+        if (isset($user['groups'])) {
+            foreach ($user['groups'] as $group) {
+                self::assertValidUserGroup($group);
+            }
+        }
+
+        if (isset($user['sub_account_ids'])) {
+            self::assertInternalType(IsType::TYPE_STRING, $user['sub_account_ids']);
+        }
+
+        foreach ($values as $key => $value) {
+            self::assertEquals($value, $user[$key]);
+        }
+    }
+
+    /**
+     * Assert that a given object is a valid user sub account.
+     * Optionally checks it against given values.
+     *
+     * @param mixed $user
+     * @param array $values
+     */
+    protected static function assertValidSubAccountUser($user, $values = [])
+    {
+        self::assertObjectStructure(
+            $user,
+            [
+                'id' => IsType::TYPE_STRING,
+                'name' => IsType::TYPE_STRING,
+                'description' => [IsType::TYPE_STRING, IsType::TYPE_NULL],
+                'cloud_name' => IsType::TYPE_STRING,
+                'api_access_keys' => IsType::TYPE_ARRAY,
+                'enabled' => IsType::TYPE_BOOL,
+                'created_at' => IsType::TYPE_STRING,
+                'custom_attributes' => [IsType::TYPE_ARRAY, IsType::TYPE_NULL]
+            ]
+        );
+
+        foreach ($user['api_access_keys'] as $api_access_keys) {
+            self::assertObjectStructure(
+                $api_access_keys,
+                [
+                    'key' => IsType::TYPE_STRING,
+                    'secret' => IsType::TYPE_STRING,
+                    'enabled' => IsType::TYPE_BOOL
+                ]
+            );
+        }
+
+        foreach ($values as $key => $value) {
+            self::assertEquals($value, $user[$key]);
+        }
+    }
+
+    /**
+     * Assert that a given object is a valid user group.
+     * Optionally checks it against given values.
+     *
+     * @param mixed $userGroup
+     * @param array $values
+     */
+    protected static function assertValidUserGroup($userGroup, $values = [])
+    {
+        self::assertObjectStructure(
+            $userGroup,
+            [
+                'id' => IsType::TYPE_STRING,
+                'name' => IsType::TYPE_STRING
+            ]
+        );
+
+        foreach ($values as $key => $value) {
+            self::assertEquals($value, $userGroup[$key]);
+        }
+    }
+
+    /**
+     * Assert that a given object is a valid user in a group.
+     * Optionally checks it against given values.
+     *
+     * @param mixed $user
+     * @param array $values
+     */
+    protected static function assertValidUserGroupUsers($user, $values = [])
+    {
+        self::assertObjectStructure(
+            $user,
+            [
+                'id' => IsType::TYPE_STRING,
+                'name' => IsType::TYPE_STRING,
+                'email' => IsType::TYPE_STRING
+            ]
+        );
+
+        foreach ($values as $key => $value) {
+            self::assertEquals($value, $user[$key]);
+        }
+    }
+
+    /**
+     * Assert that a sub account was deleted.
+     *
+     * @param $result
+     */
+    protected static function assertSubAccountDeleted($result)
+    {
+        self::assertEquals('ok', $result['message']);
+    }
+
+    /**
+     * Assert that a user was deleted.
+     *
+     * @param $result
+     */
+    protected static function assertUserDeleted($result)
+    {
+        self::assertEquals('ok', $result['message']);
+    }
+
+    /**
+     * Assert that a user group was deleted.
+     *
+     * @param $result
+     */
+    protected static function assertUserGroupDeleted($result)
+    {
+        self::assertTrue($result['ok']);
     }
 }
